serving/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
torch==2.1.0
transformers==4.40.0
onnx==1.16.0
onnxruntime-gpu==1.18.0
numpy==1.26.4
pydantic==2.7.0
requests==2.31.0
huggingface-hub==0.23.0
tritonclient[http]==2.39.0
llama-cpp-python==0.2.56